# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Four Delivery API",
        "description": "API para o sistema de delivery Four Delivery",
        "contact": {
            "email": "suporte@fourdelivery.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/admin-v1",
            "description": "Admin API V1"
        },
        {
            "url": "/api/app-v1",
            "description": "App API V1"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas as categorias",
                "operationId": "0d9f6f774f1b480b5318196b0ce0dc72",
                "responses": {
                    "200": {
                        "description": "Lista de categorias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Criar nova categoria",
                "operationId": "d566fee2a2234738bf97d6ef32889c37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id",
                                    "name",
                                    "slug",
                                    "description",
                                    "image_url"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Combos"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "combos"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Melhores combos da região"
                                    },
                                    "image_url": {
                                        "type": "string",
                                        "example": "http://example.com/image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories-with-total-items": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar categorias com total de itens e resumo de estoque",
                "operationId": "0ced9750d0a45b2e16c0dddaa817d2ff",
                "responses": {
                    "200": {
                        "description": "Lista de categorias com detalhes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories-with-company-items": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar categorias com itens da empresa",
                "operationId": "ea2d95d02dc6260c53890004796ed896",
                "responses": {
                    "200": {
                        "description": "Lista de categorias e seus itens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/companies": {
            "get": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Listar empresa(s) vinculada(s) ao usuário",
                "operationId": "310e6d7efffb131e1980d53fdd3be6d6",
                "responses": {
                    "200": {
                        "description": "Dados da empresa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Empresa"
                ],
                "summary": "Criar nova empresa",
                "operationId": "8fe07ac5bd7cbba83ee59b8633456a47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "fantasy_name",
                                    "cnpj",
                                    "email",
                                    "address",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "fantasy_name": {
                                        "type": "string"
                                    },
                                    "cnpj": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa criada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/customers": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes da empresa (paginado)",
                "operationId": "1171e97ffd82e321d87ada9c254f0324",
                "responses": {
                    "200": {
                        "description": "Lista paginada de clientes"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastrar novo cliente",
                "operationId": "8601a26c46c986ca44d13b83a75103dd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "phone",
                                    "address",
                                    "neighborhood",
                                    "city",
                                    "state",
                                    "zip_code"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "neighborhood": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "zip_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente cadastrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/items": {
            "get": {
                "tags": [
                    "Itens"
                ],
                "summary": "Listar itens da empresa (paginado)",
                "operationId": "34918d7fb0098f3716dac14cb8d062b6",
                "responses": {
                    "200": {
                        "description": "Lista paginada de itens"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Itens"
                ],
                "summary": "Criar novo item",
                "operationId": "b2b04e10cbe29048140f5c124bfb7351",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "quantity",
                                    "available",
                                    "show_in_menu"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items-all": {
            "get": {
                "tags": [
                    "Itens"
                ],
                "summary": "Listar todos os itens disponíveis (sem paginação)",
                "operationId": "e015386ada3b8682e5a90e8875864535",
                "responses": {
                    "200": {
                        "description": "Lista de itens"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/category/{categoryId}": {
            "get": {
                "tags": [
                    "Itens"
                ],
                "summary": "Listar itens por categoria",
                "operationId": "bb557e8e4bd5fb43829e29c56aeedb6c",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de itens"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items-only-available/category/{categoryId}": {
            "get": {
                "tags": [
                    "Itens"
                ],
                "summary": "Listar apenas itens disponíveis por categoria",
                "operationId": "a10e8d4618ad2188739b74d7e15bf4b6",
                "parameters": [
                    {
                        "name": "categoryId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de itens disponíveis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/items/{item}/available": {
            "patch": {
                "tags": [
                    "Itens"
                ],
                "summary": "Atualizar disponibilidade do item",
                "operationId": "6b3c03d0242fc89cf72fee2fcff6ab91",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "available"
                                ],
                                "properties": {
                                    "available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Disponibilidade atualizada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/items/{item}/show-in-menu": {
            "patch": {
                "tags": [
                    "Itens"
                ],
                "summary": "Atualizar visibilidade no menu",
                "operationId": "efd3c0eb314202f2447477cc40b27a86",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "show_in_menu"
                                ],
                                "properties": {
                                    "show_in_menu": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visibilidade atualizada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/orders/{id}/update-status": {
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Atualizar status do pedido",
                "operationId": "843d44f2311c20e03e9314b676c69adf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_id"
                                ],
                                "properties": {
                                    "status_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{id}/update-payment-status": {
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Atualizar status de pagamento",
                "operationId": "e4e355f48fc54ce8b59a2b6b98690c8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_status"
                                ],
                                "properties": {
                                    "payment_status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status de pagamento atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/me": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Mostrar dados do usuário logado",
                "operationId": "2e44f95860cf464d1eb9b77b0f964051",
                "responses": {
                    "200": {
                        "description": "Dados do usuário logado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}